+++ CHANGELOG.md
@@ -43,5 +43,6 @@ All notable changes to the Yumma CSS API will be documented in this file.
 </summary>
 
 - Use `camelCase` for exported functions instead of `UPPER_CASE`. (Internal)
+- Extend `Color` interface to include `Utility` properties (Internal)
 
 </details>